Added most frequent chart and total earnings chart

+++ app/src/main/java/com/pointlessapps/trackr/models/Salary.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 @Parcelize
-class Salary(var amount: Float = 0f, var unit: Unit = Unit.DOLLAR, var type: Type = Type.PER_HOUR) : Parcelable {
+class Salary(var amount: Double = 0.0, var unit: Unit = Unit.DOLLAR, var type: Type = Type.PER_HOUR) : Parcelable {
 	constructor(salary: Salary) : this(salary.amount, salary.unit, salary.type)
 
 	enum class Unit(@StringRes val nameResId: Int, val unit: String) {
